pyproject.toml
[tool.poetry]
name = "poweroutageanalysis"
version = "0.1.0"
description = ""
authors = ["ReadyPlayerEmma <emma@readyplayeremma.com>"]
readme = "README.md"

[project]
requires-python = ">= 3.12"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.45.1"
python-dotenv = "^1.0.1"
numpy = "^2.1.1"
rich = "^13.8.1"
pandas = "^2.2.2"
xlrd = "^2.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
ruff = "^0.4.1"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "G004"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "D104", "F401"]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff]
line-length = 120
indent-width = 4
exclude = [
    ".git",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".pyenv",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "site-packages",
    "venv",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
